Remove `imara.rs` and refactor diff viewer to use `DiffBlock` integration.
