+++ crates/diff_viewer/src/lib.rs
@@ -1,6 +1,7 @@
 mod connector;
 mod connector_builder;
-mod imara;
-mod viewer;
+mod diff_operations;
+mod rendering;
+pub mod viewer;
 
 pub use viewer::DiffViewer;
