+++ crates/diff_viewer/src/viewer.rs
@@ -1,103 +1,48 @@
-use editor::display_map::{BlockPlacement, BlockProperties, BlockStyle};
-use editor::{Editor, EditorEvent, EditorMode, RowHighlightOptions};
+use editor::{Editor, EditorEvent, EditorMode};
 use gpui::{
-    App, Background, Context, Entity, EventEmitter, FocusHandle,
-    Focusable, Hsla, IntoElement, PathBuilder, Pixels, Point as GpuiPoint, Render,
-    Subscription, Window, canvas, div, point, prelude::*, px, size,
+    App, Context, Entity, EventEmitter, FocusHandle, Focusable, IntoElement, Render, Subscription,
+    Window, div, prelude::*, px,
 };
 use language::{Buffer, Point};
 use multi_buffer::Anchor;
 use multi_buffer::MultiBuffer;
 use std::path::PathBuf;
-use std::sync::Arc;
-use theme::ActiveTheme;
 
-use crate::connector::{ConnectorCurve, ConnectorKind};
-use crate::connector_builder::build_connector_curves;
-use crate::imara::{ImaraBlockOperation, ImaraDiffAnalysis, compute_imara_diff_default};
-
-const DIFF_HIGHLIGHT_ALPHA: f32 = 0.5;
-
-fn get_diff_colors(cx: &Context<DiffViewer>) -> (Hsla, Hsla, Hsla) {
-    let theme = cx.theme();
-    let mut deleted_bg = theme.status().deleted_background;
-    deleted_bg.a = DIFF_HIGHLIGHT_ALPHA;
-    let mut created_bg = theme.status().created_background;
-    created_bg.a = DIFF_HIGHLIGHT_ALPHA;
-    let mut modified_bg = theme.status().modified_background;
-    modified_bg.a = DIFF_HIGHLIGHT_ALPHA;
-    (deleted_bg, created_bg, modified_bg)
-}
+use crate::connector::ConnectorCurve;
+use crate::connector_builder::{DiffBlock, build_connector_curves};
 
 #[derive(Clone, Copy, Debug)]
-enum PendingScroll {
+pub enum PendingScroll {
     LeftToRight { source_rows: f32 },
     RightToLeft { source_rows: f32 },
 }
 
-fn count_lines(content: &str) -> usize {
-    if content.is_empty() {
-        1
-    } else {
-        content.split('\n').count().max(1)
-    }
-}
-
-fn cubic_bezier(
-    p0: GpuiPoint<Pixels>,
-    p1: GpuiPoint<Pixels>,
-    p2: GpuiPoint<Pixels>,
-    p3: GpuiPoint<Pixels>,
-    t: f32,
-) -> GpuiPoint<Pixels> {
-    let u = 1.0 - t;
-    let tt = t * t;
-    let uu = u * u;
-    let uuu = uu * u;
-    let ttt = tt * t;
-
-    point(
-        px(uuu * f32::from(p0.x)
-            + 3.0 * uu * t * f32::from(p1.x)
-            + 3.0 * u * tt * f32::from(p2.x)
-            + ttt * f32::from(p3.x)),
-        px(uuu * f32::from(p0.y)
-            + 3.0 * uu * t * f32::from(p1.y)
-            + 3.0 * u * tt * f32::from(p2.y)
-            + ttt * f32::from(p3.y)),
-    )
-}
-
-struct DiffAdditionHighlight;
-struct DiffDeletionHighlight;
-struct DiffModificationHighlight;
-
 pub struct DiffViewer {
-    left_editor: Entity<Editor>,
-    right_editor: Entity<Editor>,
-    left_buffer: Entity<Buffer>,
-    right_buffer: Entity<Buffer>,
-    left_multibuffer: Entity<MultiBuffer>,
-    right_multibuffer: Entity<MultiBuffer>,
-    focus_handle: FocusHandle,
+    pub left_editor: Entity<Editor>,
+    pub right_editor: Entity<Editor>,
+    pub left_buffer: Entity<Buffer>,
+    pub right_buffer: Entity<Buffer>,
+    pub left_multibuffer: Entity<MultiBuffer>,
+    pub right_multibuffer: Entity<MultiBuffer>,
+    pub focus_handle: FocusHandle,
 
-    diff_analysis: Option<ImaraDiffAnalysis>,
-    connector_curves: Vec<ConnectorCurve>,
-    line_height: f32,
-    left_scroll_offset: f32,
-    right_scroll_offset: f32,
-    needs_scroll_reset: bool,
-    is_syncing_scroll: bool,
-    left_total_lines: usize,
-    right_total_lines: usize,
-    left_visible_lines: f32,
-    right_visible_lines: f32,
-    left_scroll_rows: f32,
-    right_scroll_rows: f32,
-    pending_scroll: Option<PendingScroll>,
-    _subscriptions: Vec<Subscription>,
-    left_crushed_blocks: Vec<editor::display_map::CustomBlockId>,
-    right_crushed_blocks: Vec<editor::display_map::CustomBlockId>,
+    pub diff_blocks: Vec<DiffBlock>,
+    pub connector_curves: Vec<ConnectorCurve>,
+    pub line_height: f32,
+    pub left_scroll_offset: f32,
+    pub right_scroll_offset: f32,
+    pub needs_scroll_reset: bool,
+    pub is_syncing_scroll: bool,
+    pub left_total_lines: usize,
+    pub right_total_lines: usize,
+    pub left_visible_lines: f32,
+    pub right_visible_lines: f32,
+    pub left_scroll_rows: f32,
+    pub right_scroll_rows: f32,
+    pub pending_scroll: Option<PendingScroll>,
+    pub _subscriptions: Vec<Subscription>,
+    pub left_crushed_blocks: Vec<editor::display_map::CustomBlockId>,
+    pub right_crushed_blocks: Vec<editor::display_map::CustomBlockId>,
 }
 
 impl EventEmitter<()> for DiffViewer {}
@@ -129,29 +74,14 @@ impl DiffViewer {
         cx.notify();
     }
 
-    fn left_line_to_anchor(&self, line: u32, cx: &Context<Self>) -> Anchor {
+    pub fn left_line_to_anchor(&self, line: u32, cx: &Context<Self>) -> Anchor {
         let snapshot = self.left_multibuffer.read(cx).snapshot(cx);
         snapshot.anchor_before(Point::new(line, 0))
     }
 
-    fn right_line_to_anchor(&self, line: u32, cx: &Context<Self>) -> Anchor {
+    pub fn right_line_to_anchor(&self, line: u32, cx: &Context<Self>) -> Anchor {
         let snapshot = self.right_multibuffer.read(cx).snapshot(cx);
         snapshot.anchor_before(Point::new(line, 0))
-    }
-
-    fn create_crushed_block_properties(
-        &self,
-        anchor: Anchor,
-        color: Hsla,
-        _cx: &Context<Self>,
-    ) -> BlockProperties<Anchor> {
-        BlockProperties {
-            placement: BlockPlacement::Replace(anchor..=anchor),
-            height: Some(2),
-            style: BlockStyle::Fixed,
-            render: Arc::new(move |_| div().absolute().w_full().h(px(2.0)).bg(color).into_any()),
-            priority: 0,
-        }
     }
 
     pub fn new(
@@ -199,7 +129,7 @@ impl DiffViewer {
                 window,
                 cx,
             );
-            editor.set_read_only(true);
+            editor.set_read_only(false);
             editor.set_show_gutter(true, cx);
             editor.set_show_scrollbars(true, cx);
             editor
@@ -215,7 +145,7 @@ impl DiffViewer {
 
         let default_visible_lines = viewport_height / line_height;
 
-        let viewer = Self {
+        Self {
             left_editor,
             right_editor,
             left_buffer,
@@ -223,7 +153,7 @@ impl DiffViewer {
             left_multibuffer,
             right_multibuffer,
             focus_handle: cx.focus_handle(),
-            diff_analysis: None,
+            diff_blocks: Vec::new(),
             connector_curves: Vec::new(),
             line_height,
             left_scroll_offset: 0.0,
@@ -240,20 +170,17 @@ impl DiffViewer {
             _subscriptions: Vec::new(),
             left_crushed_blocks: Vec::new(),
             right_crushed_blocks: Vec::new(),
-        };
-
-        viewer
+        }
     }
 
     pub fn initialize(&mut self, _window: &mut Window, cx: &mut Context<Self>) {
         let left_subscription = cx.subscribe(
             &self.left_editor,
-            |this: &mut DiffViewer, _editor, event: &EditorEvent, cx| {
-                if let EditorEvent::ScrollPositionChanged {
+            |this: &mut DiffViewer, _editor, event: &EditorEvent, cx| match event {
+                EditorEvent::ScrollPositionChanged {
                     autoscroll: _,
                     local: _,
-                } = event
-                {
+                } => {
                     if this.is_syncing_scroll {
                         return;
                     }
@@ -268,17 +195,20 @@ impl DiffViewer {
                         this.request_sync_from_left(rows as f32, cx);
                     }
                 }
+                EditorEvent::BufferEdited | EditorEvent::Edited { .. } => {
+                    this.refresh_diff_on_content_change(cx);
+                }
+                _ => {}
             },
         );
 
         let right_subscription = cx.subscribe(
             &self.right_editor,
-            |this: &mut DiffViewer, _editor, event: &EditorEvent, cx| {
-                if let EditorEvent::ScrollPositionChanged {
+            |this: &mut DiffViewer, _editor, event: &EditorEvent, cx| match event {
+                EditorEvent::ScrollPositionChanged {
                     autoscroll: _,
                     local: _,
-                } = event
-                {
+                } => {
                     if this.is_syncing_scroll {
                         return;
                     }
@@ -293,6 +223,10 @@ impl DiffViewer {
                         this.request_sync_from_right(rows as f32, cx);
                     }
                 }
+                EditorEvent::BufferEdited | EditorEvent::Edited { .. } => {
+                    this.refresh_diff_on_content_change(cx);
+                }
+                _ => {}
             },
         );
 
@@ -335,16 +269,13 @@ impl DiffViewer {
             buffer.edit([(0..buffer.len(), right_content.clone())], None, cx);
         });
 
+        use crate::diff_operations::count_lines;
         self.left_total_lines = count_lines(&left_content);
         self.right_total_lines = count_lines(&right_content);
 
-        let analysis = compute_imara_diff_default(&left_content, &right_content);
-
-        self.diff_analysis = Some(analysis.clone());
-
-        self.connector_curves = build_connector_curves(&analysis);
-
-        self.apply_diff_highlights(&analysis, cx);
+        self.diff_blocks = self.extract_diff_blocks(cx);
+        self.connector_curves = build_connector_curves(&self.diff_blocks);
+        self.apply_diff_highlights(cx);
 
         self.pending_scroll = None;
         self.needs_scroll_reset = true;
@@ -356,810 +287,86 @@ impl DiffViewer {
         cx.notify();
     }
 
-    fn apply_diff_highlights(&mut self, analysis: &ImaraDiffAnalysis, cx: &mut Context<Self>) {
-        let (deleted_bg, created_bg, modified_bg) = get_diff_colors(cx);
-
-        self.left_editor.update(cx, |editor, _cx| {
-            editor.clear_row_highlights::<DiffDeletionHighlight>();
-            editor.clear_row_highlights::<DiffModificationHighlight>();
+    pub fn set_right_buffer(
+        &mut self,
+        buffer: Entity<Buffer>,
+        left_content: String,
+        window: &mut Window,
+        cx: &mut Context<Self>,
+    ) {
+        self.left_buffer.update(cx, |buf, cx| {
+            buf.edit([(0..buf.len(), left_content.clone())], None, cx);
         });
 
-        self.right_editor.update(cx, |editor, _cx| {
-            editor.clear_row_highlights::<DiffAdditionHighlight>();
-            editor.clear_row_highlights::<DiffModificationHighlight>();
-        });
-
-        for block in &analysis.blocks {
-            match block.operation {
-                ImaraBlockOperation::Delete => {
-                    if !block.left_range.is_empty() {
-                        self.left_editor.update(cx, |editor, cx| {
-                            let start_row = block.left_range.start as u32;
-                            let end_row = block
-                                .left_range
-                                .end
-                                .saturating_sub(1)
-                                .max(block.left_range.start)
-                                as u32;
-
-                            let buffer = editor.buffer().read(cx);
-                            let snapshot = buffer.snapshot(cx);
-
-                            let actual_end_row = end_row.min(snapshot.max_row().0);
-                            let start_anchor = snapshot.anchor_before(Point::new(start_row, 0));
-                            let end_anchor =
-                                snapshot.anchor_before(Point::new(actual_end_row + 1, 0));
-
-                            editor.highlight_rows::<DiffDeletionHighlight>(
-                                start_anchor..end_anchor,
-                                deleted_bg,
-                                RowHighlightOptions {
-                                    autoscroll: false,
-                                    include_gutter: true,
-                                },
-                                cx,
-                            );
-                        });
-                    }
-                }
-                ImaraBlockOperation::Insert => {
-                    if !block.right_range.is_empty() {
-                        self.right_editor.update(cx, |editor, cx| {
-                            let start_row = block.right_range.start as u32;
-                            let end_row = block
-                                .right_range
-                                .end
-                                .saturating_sub(1)
-                                .max(block.right_range.start)
-                                as u32;
-
-                            let buffer = editor.buffer().read(cx);
-                            let snapshot = buffer.snapshot(cx);
-
-                            let actual_end_row = end_row.min(snapshot.max_row().0);
-                            let start_anchor = snapshot.anchor_before(Point::new(start_row, 0));
-                            let end_anchor =
-                                snapshot.anchor_before(Point::new(actual_end_row + 1, 0));
-
-                            editor.highlight_rows::<DiffAdditionHighlight>(
-                                start_anchor..end_anchor,
-                                created_bg,
-                                RowHighlightOptions {
-                                    autoscroll: false,
-                                    include_gutter: true,
-                                },
-                                cx,
-                            );
-                        });
-                    }
-                }
-                ImaraBlockOperation::Modify => {
-                    if !block.left_range.is_empty() {
-                        self.left_editor.update(cx, |editor, cx| {
-                            let start_row = block.left_range.start as u32;
-                            let end_row = block
-                                .left_range
-                                .end
-                                .saturating_sub(1)
-                                .max(block.left_range.start)
-                                as u32;
-
-                            let buffer = editor.buffer().read(cx);
-                            let snapshot = buffer.snapshot(cx);
-
-                            let actual_end_row = end_row.min(snapshot.max_row().0);
-                            let start_anchor = snapshot.anchor_before(Point::new(start_row, 0));
-                            let end_anchor =
-                                snapshot.anchor_before(Point::new(actual_end_row + 1, 0));
-
-                            editor.highlight_rows::<DiffModificationHighlight>(
-                                start_anchor..end_anchor,
-                                modified_bg,
-                                RowHighlightOptions {
-                                    autoscroll: false,
-                                    include_gutter: true,
-                                },
-                                cx,
-                            );
-                        });
-                    }
-
-                    if !block.right_range.is_empty() {
-                        self.right_editor.update(cx, |editor, cx| {
-                            let start_row = block.right_range.start as u32;
-                            let end_row = block
-                                .right_range
-                                .end
-                                .saturating_sub(1)
-                                .max(block.right_range.start)
-                                as u32;
-
-                            let buffer = editor.buffer().read(cx);
-                            let snapshot = buffer.snapshot(cx);
-
-                            let actual_end_row = end_row.min(snapshot.max_row().0);
-                            let start_anchor = snapshot.anchor_before(Point::new(start_row, 0));
-                            let end_anchor =
-                                snapshot.anchor_before(Point::new(actual_end_row + 1, 0));
-
-                            editor.highlight_rows::<DiffModificationHighlight>(
-                                start_anchor..end_anchor,
-                                modified_bg,
-                                RowHighlightOptions {
-                                    autoscroll: false,
-                                    include_gutter: true,
-                                },
-                                cx,
-                            );
-                        });
-                    }
-                }
-            }
-        }
-    }
-
-    fn render_left_crushed_blocks(&self, cx: &Context<Self>) -> impl IntoElement {
-        let curves = self.connector_curves.clone();
-        let left_editor = self.left_editor.clone();
-        let fallback_line_height = self.line_height;
-
-        let (_deleted_bg, created_bg, _modified_bg) = get_diff_colors(cx);
-
-        #[derive(Clone)]
-        struct LeftCrushedCanvasData {
-            curves: Vec<ConnectorCurve>,
-            line_height: f32,
-            left_scroll_pixels: f32,
-            left_top_origin: f32,
-            created_color: Hsla,
-        }
-
-        canvas(
-            move |bounds, window, cx| {
-                let (left_line_height, left_scroll_pixels, left_bounds) =
-                    left_editor.update(cx, |editor, cx| {
-                        let line_height = editor
-                            .style()
-                            .map(|style| {
-                                f32::from(style.text.line_height_in_pixels(window.rem_size()))
-                            })
-                            .unwrap_or(fallback_line_height);
-
-                        let scroll_rows = editor.scroll_position(cx).y;
-                        let scroll_pixels = (scroll_rows as f32) * line_height;
-                        let bounds = editor.last_bounds().cloned();
-
-                        (line_height, scroll_pixels, bounds)
-                    });
-
-                let left_top_origin = left_bounds
-                    .as_ref()
-                    .map(|b| f32::from(b.origin.y))
-                    .unwrap_or(f32::from(bounds.origin.y));
-
-                LeftCrushedCanvasData {
-                    curves,
-                    line_height: left_line_height,
-                    left_scroll_pixels,
-                    left_top_origin,
-                    created_color: created_bg,
-                }
-            },
-            move |bounds, data, window, _cx| {
-                if data.curves.is_empty() {
-                    return;
-                }
-
-                let _header_height = data.left_top_origin - f32::from(bounds.origin.y);
-                let crushed_thickness = 2.0;
-                let minimal_block_height = 2.0;
-                let mut deleted_lines_above = 0usize;
-
-                for curve in &data.curves {
-                    let left_len = curve.left_end.saturating_sub(curve.left_start) + 1;
-                    let right_len = curve.right_end.saturating_sub(curve.right_start) + 1;
-
-                    if curve.left_crushed {
-                    } else if curve.right_crushed {
-                        deleted_lines_above += left_len;
-                    } else {
-                        if right_len < left_len {
-                            deleted_lines_above += left_len - right_len;
-                        }
-                    }
-
-                    if curve.left_crushed {
-                        let left_offset_rows = deleted_lines_above as f32;
-                        let left_row = curve.focus_line as f32 + left_offset_rows;
-                        let left_y = (left_row * data.line_height) - data.left_scroll_pixels;
-                        let left_bottom = left_y + minimal_block_height;
-
-                        let left_absolute_top = data.left_top_origin + left_y;
-                        let left_absolute_bottom = data.left_top_origin + left_bottom;
-
-                        let y_center = (left_absolute_top + left_absolute_bottom) * 0.5;
-                        let top = y_center - crushed_thickness / 2.0;
-                        let bottom = top + crushed_thickness;
-
-                        if bottom > data.left_top_origin {
-                            let clipped_top = top.max(data.left_top_origin);
-                            let clipped_bottom = bottom.max(data.left_top_origin);
-
-                            let mut builder = PathBuilder::fill();
-                            builder.move_to(point(px(f32::from(bounds.origin.x)), px(clipped_top)));
-                            builder.line_to(point(
-                                px(f32::from(bounds.origin.x) + f32::from(bounds.size.width)),
-                                px(clipped_top),
-                            ));
-                            builder.line_to(point(
-                                px(f32::from(bounds.origin.x) + f32::from(bounds.size.width)),
-                                px(clipped_bottom),
-                            ));
-                            builder.line_to(point(px(f32::from(bounds.origin.x)), px(clipped_bottom)));
-                            builder.close();
-
-                            if let Ok(path) = builder.build() {
-                                let background: Background = data.created_color.into();
-                                window.paint_path(path, background);
-                            }
-                        }
-                    }
-                }
-            },
-        )
-        .size_full()
-    }
-
-
-    fn render_right_crushed_blocks(&self, cx: &Context<Self>) -> impl IntoElement {
-        let curves = self.connector_curves.clone();
-        let right_editor = self.right_editor.clone();
-        let fallback_line_height = self.line_height;
-
-        let (deleted_bg, _created_bg, _modified_bg) = get_diff_colors(cx);
-
-        #[derive(Clone)]
-        struct RightCrushedCanvasData {
-            curves: Vec<ConnectorCurve>,
-            line_height: f32,
-            right_scroll_pixels: f32,
-            right_top_origin: f32,
-            deleted_color: Hsla,
-        }
-
-        canvas(
-            move |bounds, window, cx| {
-                let (right_line_height, right_scroll_pixels, right_bounds) =
-                    right_editor.update(cx, |editor, cx| {
-                        let line_height = editor
-                            .style()
-                            .map(|style| {
-                                f32::from(style.text.line_height_in_pixels(window.rem_size()))
-                            })
-                            .unwrap_or(fallback_line_height);
-
-                        let scroll_rows = editor.scroll_position(cx).y;
-                        let scroll_pixels = (scroll_rows as f32) * line_height;
-                        let bounds = editor.last_bounds().cloned();
-
-                        (line_height, scroll_pixels, bounds)
-                    });
-
-                let right_top_origin = right_bounds
-                    .as_ref()
-                    .map(|b| f32::from(b.origin.y))
-                    .unwrap_or(f32::from(bounds.origin.y));
-
-                RightCrushedCanvasData {
-                    curves,
-                    line_height: right_line_height,
-                    right_scroll_pixels,
-                    right_top_origin,
-                    deleted_color: deleted_bg,
-                }
-            },
-            move |bounds, data, window, _cx| {
-                if data.curves.is_empty() {
-                    return;
-                }
-
-                let crushed_thickness = 2.0;
-                let minimal_block_height = 2.0;
-                let mut inserted_lines_above = 0usize;
-
-                for curve in &data.curves {
-                    let left_len = curve.left_end.saturating_sub(curve.left_start) + 1;
-                    let right_len = curve.right_end.saturating_sub(curve.right_start) + 1;
-
-                    if curve.left_crushed {
-                        inserted_lines_above += right_len;
-                    } else if curve.right_crushed {
-                    } else {
-                        if left_len < right_len {
-                            inserted_lines_above += right_len - left_len;
-                        }
-                    }
-
-                    if curve.right_crushed {
-                        let right_offset_rows = inserted_lines_above as f32;
-                        let right_row = curve.focus_line as f32 + right_offset_rows;
-                        let right_y = (right_row * data.line_height) - data.right_scroll_pixels;
-                        let right_bottom = right_y + minimal_block_height;
-
-                        let right_absolute_top = data.right_top_origin + right_y;
-                        let right_absolute_bottom = data.right_top_origin + right_bottom;
-
-                        let y_center = (right_absolute_top + right_absolute_bottom) * 0.5;
-                        let top = y_center - crushed_thickness / 2.0;
-                        let bottom = top + crushed_thickness;
-
-                        if bottom > data.right_top_origin {
-                            let clipped_top = top.max(data.right_top_origin);
-                            let clipped_bottom = bottom.max(data.right_top_origin);
-
-                            let mut builder = PathBuilder::fill();
-                            builder.move_to(point(px(f32::from(bounds.origin.x)), px(clipped_top)));
-                            builder.line_to(point(
-                                px(f32::from(bounds.origin.x) + f32::from(bounds.size.width)),
-                                px(clipped_top),
-                            ));
-                            builder.line_to(point(
-                                px(f32::from(bounds.origin.x) + f32::from(bounds.size.width)),
-                                px(clipped_bottom),
-                            ));
-                            builder.line_to(point(px(f32::from(bounds.origin.x)), px(clipped_bottom)));
-                            builder.close();
-
-                            if let Ok(path) = builder.build() {
-                                let background: Background = data.deleted_color.into();
-                                window.paint_path(path, background);
-                            }
-                        }
-                    }
-                }
-            },
-        )
-        .size_full()
-    }
-
-    fn render_connectors(&self, cx: &Context<Self>) -> impl IntoElement {
-        let curves = self.connector_curves.clone();
-        let left_editor = self.left_editor.clone();
-        let right_editor = self.right_editor.clone();
-        let fallback_line_height = self.line_height;
-
-        let (deleted_bg, created_bg, modified_bg) = get_diff_colors(cx);
-
-        #[derive(Clone)]
-        struct ConnectorCanvasData {
-            curves: Vec<ConnectorCurve>,
-            line_height: f32,
-            left_scroll_pixels: f32,
-            right_scroll_pixels: f32,
-            left_top_origin: f32,
-            right_top_origin: f32,
-            left_bounds: Option<gpui::Bounds<Pixels>>,
-            right_bounds: Option<gpui::Bounds<Pixels>>,
-            created_bg: Hsla,
-            deleted_bg: Hsla,
-            modified_bg: Hsla,
-        }
-
-        canvas(
-            move |bounds, window, cx| {
-                let (left_line_height, left_scroll_pixels, left_bounds) =
-                    left_editor.update(cx, |editor, cx| {
-                        let line_height = editor
-                            .style()
-                            .map(|style| {
-                                f32::from(style.text.line_height_in_pixels(window.rem_size()))
-                            })
-                            .unwrap_or(fallback_line_height);
-
-                        let scroll_rows = editor.scroll_position(cx).y;
-                        let scroll_pixels = (scroll_rows as f32) * line_height;
-                        let bounds = editor.last_bounds().cloned();
-
-                        (line_height, scroll_pixels, bounds)
-                    });
-
-                let (_right_line_height, right_scroll_pixels, right_bounds) =
-                    right_editor.update(cx, |editor, cx| {
-                        let line_height = editor
-                            .style()
-                            .map(|style| {
-                                f32::from(style.text.line_height_in_pixels(window.rem_size()))
-                            })
-                            .unwrap_or(fallback_line_height);
-
-                        let scroll_rows = editor.scroll_position(cx).y;
-                        let scroll_pixels = (scroll_rows as f32) * line_height;
-                        let bounds = editor.last_bounds().cloned();
-
-                        (line_height, scroll_pixels, bounds)
-                    });
-
-                let line_height = left_line_height;
-                let left_top_origin = left_bounds
-                    .as_ref()
-                    .map(|b| f32::from(b.origin.y))
-                    .unwrap_or(f32::from(bounds.origin.y));
-                let right_top_origin = right_bounds
-                    .as_ref()
-                    .map(|b| f32::from(b.origin.y))
-                    .unwrap_or(f32::from(bounds.origin.y));
-
-                ConnectorCanvasData {
-                    curves,
-                    line_height,
-                    left_scroll_pixels,
-                    right_scroll_pixels,
-                    left_top_origin,
-                    right_top_origin,
-                    left_bounds,
-                    right_bounds,
-                    created_bg,
-                    deleted_bg,
-                    modified_bg,
-                }
-            },
-            move |bounds, data, window, _cx| {
-                if data.curves.is_empty() {
-                    return;
-                }
-
-                let gutter_width = f32::from(bounds.size.width);
-
-                let header_height = data.left_top_origin - f32::from(bounds.origin.y);
-                let viewport_top = header_height;
-                let viewport_bottom = f32::from(bounds.size.height);
-
-                let left_offset = data.left_top_origin - f32::from(bounds.origin.y);
-                let right_offset = data.right_top_origin - f32::from(bounds.origin.y);
-
-                let minimal_block_height = 2.0;
-                let mut inserted_lines_above = 0usize;
-                let mut deleted_lines_above = 0usize;
-
-                for curve in &data.curves {
-                    let is_left_empty = curve.left_crushed;
-                    let is_right_empty = curve.right_crushed;
-
-                    let left_offset_rows = if is_left_empty {
-                        deleted_lines_above as f32
-                    } else {
-                        0.0
-                    };
-
-                    let right_offset_rows = if is_right_empty {
-                        inserted_lines_above as f32
-                    } else {
-                        0.0
-                    };
-
-                    let left_len = curve.left_end.saturating_sub(curve.left_start) + 1;
-                    let right_len = curve.right_end.saturating_sub(curve.right_start) + 1;
-
-                    if curve.left_crushed {
-                        inserted_lines_above += right_len;
-                    } else if curve.right_crushed {
-                        deleted_lines_above += left_len;
-                    } else {
-                        if left_len < right_len {
-                            inserted_lines_above += right_len - left_len;
-                        } else if right_len < left_len {
-                            deleted_lines_above += left_len - right_len;
-                        }
-                    }
-
-                    let left_row = if is_left_empty {
-                        curve.focus_line as f32 + left_offset_rows
-                    } else {
-                        curve.left_start as f32
-                    };
-
-                    let right_row = if is_right_empty {
-                        curve.focus_line as f32 + right_offset_rows
-                    } else {
-                        curve.right_start as f32
-                    };
-
-                    let left_y = (left_row * data.line_height) - data.left_scroll_pixels;
-                    let right_y = (right_row * data.line_height) - data.right_scroll_pixels;
-
-                    let left_bottom = if is_left_empty {
-                        left_y + minimal_block_height
-                    } else {
-                        ((curve.left_end as f32 + 1.0) * data.line_height - data.left_scroll_pixels)
-                            .max(left_y + minimal_block_height)
-                    };
-
-                    let right_bottom = if is_right_empty {
-                        right_y + minimal_block_height
-                    } else {
-                        ((curve.right_end as f32 + 1.0) * data.line_height
-                            - data.right_scroll_pixels)
-                            .max(right_y + minimal_block_height)
-                    };
-
-                    let left_top = left_y;
-                    let right_top = right_y;
-
-                    let left_absolute_top = data.left_top_origin + left_top;
-                    let left_absolute_bottom = data.left_top_origin + left_bottom;
-                    let right_absolute_top = data.right_top_origin + right_top;
-                    let right_absolute_bottom = data.right_top_origin + right_bottom;
-
-                    let adjusted_left_top = left_top + left_offset;
-                    let adjusted_left_bottom = left_bottom + left_offset;
-                    let adjusted_right_top = right_top + right_offset;
-                    let adjusted_right_bottom = right_bottom + right_offset;
-
-                    let connector_height = (adjusted_left_bottom - adjusted_left_top)
-                        .max(adjusted_right_bottom - adjusted_right_top);
-                    let base_control_offset = gutter_width * 0.25;
-                    let reference_line_height = data.line_height.max(1.0);
-                    let control_offset = if connector_height < reference_line_height * 2.0 {
-                        base_control_offset
-                            * (connector_height / (reference_line_height * 2.0)).max(0.3)
-                    } else {
-                        base_control_offset
-                    };
-
-                    let connector_top = adjusted_left_top.min(adjusted_right_top);
-                    let connector_bottom = adjusted_left_bottom.max(adjusted_right_bottom);
-
-                    let base_color = match curve.kind {
-                        ConnectorKind::Insert => data.created_bg,
-                        ConnectorKind::Delete => data.deleted_bg,
-                        ConnectorKind::Modify => data.modified_bg,
-                    };
-
-                    let is_visible =
-                        connector_bottom >= viewport_top && connector_top <= viewport_bottom;
-
-                    if is_visible {
-                        Self::draw_crushed_indicator(
-                            window,
-                            &bounds,
-                            data.left_bounds.as_ref(),
-                            data.right_bounds.as_ref(),
-                            is_left_empty,
-                            is_right_empty,
-                            left_absolute_top,
-                            right_absolute_top,
-                            left_absolute_bottom,
-                            right_absolute_bottom,
-                            gutter_width,
-                            base_color,
-                        );
-                    }
-
-                    let thickness_multiplier = match curve.kind {
-                        ConnectorKind::Modify => {
-                            let line_count = ((curve.left_end - curve.left_start)
-                                .max(curve.right_end - curve.right_start))
-                                as u32;
-                            if line_count > 5 {
-                                1.3
-                            } else if line_count > 1 {
-                                1.15
-                            } else {
-                                1.0
-                            }
-                        }
-                        _ => 1.0,
-                    };
-
-                    let _clipped_left_top = adjusted_left_top.max(header_height);
-                    let _clipped_right_top = adjusted_right_top.max(header_height);
-
-                    let has_left_visible = adjusted_left_bottom > header_height
-                        && adjusted_left_top < adjusted_left_bottom;
-                    let has_right_visible = adjusted_right_bottom > header_height
-                        && adjusted_right_top < adjusted_right_bottom;
-
-                    if is_visible && (has_left_visible || has_right_visible) {
-                        Self::draw_connector_ribbon(
-                            window,
-                            &bounds,
-                            adjusted_left_top,
-                            adjusted_left_bottom,
-                            adjusted_right_top,
-                            adjusted_right_bottom,
-                            control_offset,
-                            base_color,
-                            thickness_multiplier,
-                            header_height,
-                        );
-                    }
-                }
-            },
-        )
-        .size_full()
-    }
-
-    fn draw_crushed_indicator(
-        window: &mut Window,
-        gutter_bounds: &gpui::Bounds<Pixels>,
-        _left_bounds: Option<&gpui::Bounds<Pixels>>,
-        _right_bounds: Option<&gpui::Bounds<Pixels>>,
-        left_crushed: bool,
-        right_crushed: bool,
-        left_top: f32,
-        right_top: f32,
-        left_bottom: f32,
-        right_bottom: f32,
-        _gutter_width: f32,
-        color: gpui::Hsla,
-    ) {
-        let crushed_thickness = 2.0;
-
-        if left_crushed && right_crushed {
-            let y_center = ((left_top + left_bottom) + (right_top + right_bottom)) * 0.25;
-            let top = f32::from(gutter_bounds.origin.y) + y_center - crushed_thickness / 2.0;
-            let bottom = top + crushed_thickness;
-            let left = f32::from(gutter_bounds.origin.x);
-            let right = f32::from(gutter_bounds.origin.x) + f32::from(gutter_bounds.size.width);
-            let mut builder = PathBuilder::fill();
-            builder.move_to(point(px(left), px(top)));
-            builder.line_to(point(px(right), px(top)));
-            builder.line_to(point(px(right), px(bottom)));
-            builder.line_to(point(px(left), px(bottom)));
-            builder.close();
-
-            if let Ok(path) = builder.build() {
-                let background: Background = color.into();
-                window.paint_path(path, background);
-            }
-        }
-    }
-
-    fn draw_connector_ribbon(
-        window: &mut Window,
-        bounds: &gpui::Bounds<Pixels>,
-        left_top: f32,
-        left_bottom: f32,
-        right_top: f32,
-        right_bottom: f32,
-        control_offset: f32,
-        color: gpui::Hsla,
-        thickness_multiplier: f32,
-        header_height: f32,
-    ) {
-        let _base_thickness = 6.0 * thickness_multiplier;
-        let segments = 48;
-
-        let mut builder = PathBuilder::fill();
-
-        for i in 0..=segments {
-            let t = i as f32 / segments as f32;
-            let top_point = cubic_bezier(
-                point(
-                    px(f32::from(bounds.origin.x)),
-                    px(f32::from(bounds.origin.y) + left_top),
-                ),
-                point(
-                    px(f32::from(bounds.origin.x) + control_offset),
-                    px(f32::from(bounds.origin.y) + left_top),
-                ),
-                point(
-                    px(f32::from(bounds.origin.x) + f32::from(bounds.size.width) - control_offset),
-                    px(f32::from(bounds.origin.y) + right_top),
-                ),
-                point(
-                    px(f32::from(bounds.origin.x) + f32::from(bounds.size.width)),
-                    px(f32::from(bounds.origin.y) + right_top),
-                ),
-                t,
-            );
-            if i == 0 {
-                builder.move_to(top_point);
-            } else {
-                builder.line_to(top_point);
-            }
-        }
-
-        for i in (0..=segments).rev() {
-            let t = i as f32 / segments as f32;
-            let bottom_point = cubic_bezier(
-                point(
-                    px(f32::from(bounds.origin.x)),
-                    px(f32::from(bounds.origin.y) + left_bottom),
-                ),
-                point(
-                    px(f32::from(bounds.origin.x) + control_offset),
-                    px(f32::from(bounds.origin.y) + left_bottom),
-                ),
-                point(
-                    px(f32::from(bounds.origin.x) + f32::from(bounds.size.width) - control_offset),
-                    px(f32::from(bounds.origin.y) + right_bottom),
-                ),
-                point(
-                    px(f32::from(bounds.origin.x) + f32::from(bounds.size.width)),
-                    px(f32::from(bounds.origin.y) + right_bottom),
-                ),
-                t,
-            );
-            builder.line_to(bottom_point);
-        }
-
-        if let Ok(path) = builder.build() {
-            let clip_top = f32::from(bounds.origin.y) + header_height;
-            let clip_bounds = gpui::Bounds {
-                origin: point(px(f32::from(bounds.origin.x)), px(clip_top)),
-                size: size(
-                    bounds.size.width,
-                    px(f32::from(bounds.size.height) - header_height),
-                ),
-            };
-
-            window.with_content_mask(
-                Some(gpui::ContentMask {
-                    bounds: clip_bounds,
-                }),
-                |window| {
-                    let background: Background = color.into();
-                    window.paint_path(path, background);
+        self.right_buffer = buffer.clone();
+        self.right_multibuffer = cx.new(|cx| MultiBuffer::singleton(buffer.clone(), cx));
+
+        self.right_editor = cx.new(|cx| {
+            let mut editor = Editor::new(
+                EditorMode::Full {
+                    scale_ui_elements_with_buffer_font_size: false,
+                    show_active_line_background: false,
+                    sized_by_content: false,
                 },
+                self.right_multibuffer.clone(),
+                None,
+                window,
+                cx,
             );
-        }
-    }
+            editor.set_read_only(false);
+            editor.set_show_gutter(true, cx);
+            editor.set_show_scrollbars(true, cx);
+            editor
+        });
 
-    fn update_crushed_blocks(&mut self, cx: &mut Context<Self>) {
-        let (deleted_bg, created_bg, _modified_bg) = get_diff_colors(cx);
+        let right_subscription = cx.subscribe(
+            &self.right_editor,
+            |this: &mut DiffViewer, _editor, event: &EditorEvent, cx| match event {
+                EditorEvent::ScrollPositionChanged {
+                    autoscroll: _,
+                    local: _,
+                } => {
+                    if this.is_syncing_scroll {
+                        return;
+                    }
 
-        if !self.left_crushed_blocks.is_empty() {
-            self.left_editor.update(cx, |editor, cx| {
-                editor.remove_blocks(
-                    self.left_crushed_blocks.clone().into_iter().collect(),
-                    None,
-                    cx,
-                );
-            });
-            self.left_crushed_blocks.clear();
-        }
+                    let rows = this
+                        .right_editor
+                        .update(cx, |editor, cx| editor.scroll_position(cx).y);
 
-        if !self.right_crushed_blocks.is_empty() {
-            self.right_editor.update(cx, |editor, cx| {
-                editor.remove_blocks(
-                    self.right_crushed_blocks.clone().into_iter().collect(),
-                    None,
-                    cx,
-                );
-            });
-            self.right_crushed_blocks.clear();
-        }
+                    if (rows as f32 - this.right_scroll_rows).abs() > f32::EPSILON {
+                        this.right_scroll_rows = rows as f32;
+                        this.right_scroll_offset = (rows as f32) * this.line_height;
+                        this.request_sync_from_right(rows as f32, cx);
+                    }
+                }
+                EditorEvent::BufferEdited | EditorEvent::Edited { .. } => {
+                    this.refresh_diff_on_content_change(cx);
+                }
+                _ => {}
+            },
+        );
 
-        let mut left_crushed_positions = Vec::new();
-        let mut right_crushed_positions = Vec::new();
+        self._subscriptions.truncate(1);
+        self._subscriptions.push(right_subscription);
 
-        for curve in &self.connector_curves {
-            if curve.left_crushed {
-                left_crushed_positions.push(curve.focus_line);
-            }
-            if curve.right_crushed {
-                right_crushed_positions.push(curve.focus_line);
-            }
-        }
+        let right_content = buffer.read(cx).text();
+        use crate::diff_operations::count_lines;
+        self.left_total_lines = count_lines(&left_content);
+        self.right_total_lines = count_lines(&right_content);
 
-        for line in left_crushed_positions {
-            let anchor = self.left_line_to_anchor(line as u32, cx);
-            let block_props = self.create_crushed_block_properties(anchor, created_bg, cx);
-            let block_ids = self.left_editor.update(cx, |editor, cx| {
-                editor.insert_blocks([block_props], None, cx)
-            });
-            self.left_crushed_blocks.extend(block_ids);
-        }
+        self.diff_blocks = self.extract_diff_blocks(cx);
+        self.connector_curves = build_connector_curves(&self.diff_blocks);
+        self.apply_diff_highlights(cx);
 
-        for line in right_crushed_positions {
-            let anchor = self.right_line_to_anchor(line as u32, cx);
-            let block_props = self.create_crushed_block_properties(anchor, deleted_bg, cx);
-            let block_ids = self.right_editor.update(cx, |editor, cx| {
-                editor.insert_blocks([block_props], None, cx)
-            });
-            self.right_crushed_blocks.extend(block_ids);
-        }
+        self.pending_scroll = None;
+        self.needs_scroll_reset = true;
+        self.left_scroll_offset = 0.0;
+        self.right_scroll_offset = 0.0;
+        self.left_scroll_rows = 0.0;
+        self.right_scroll_rows = 0.0;
+
+        cx.notify();
     }
 }
 
@@ -1171,6 +378,8 @@ impl Focusable for DiffViewer {
 
 impl Render for DiffViewer {
     fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
+        use theme::ActiveTheme;
+
         if let Some(visible) = self.left_editor.read(cx).visible_line_count() {
             self.left_visible_lines = visible as f32;
         }
@@ -1293,9 +502,35 @@ impl Render for DiffViewer {
                                                     .absolute()
                                                     .top_0()
                                                     .left_0()
-                                            .right_0()
-                                            .bottom_0()
+                                                    .right_0()
+                                                    .bottom_0()
+                                                    .on_mouse_move(cx.listener(
+                                                        |this, event, _, cx| {
+                                                            this.handle_mouse_move(event, cx);
+                                                        },
+                                                    )),
+                                            )
+                                            .child(
+                                                div()
+                                                    .absolute()
+                                                    .top_0()
+                                                    .left_0()
+                                                    .right_0()
+                                                    .bottom_0()
                                                     .child(self.render_left_crushed_blocks(cx)),
+                                            )
+                                            .child(
+                                                div()
+                                                    .absolute()
+                                                    .top_0()
+                                                    .left_0()
+                                                    .right_0()
+                                                    .bottom_0()
+                                                    .children(
+                                                        self.render_left_editor_revert_buttons(
+                                                            window, cx,
+                                                        ),
+                                                    ),
                                             ),
                                     ),
                             ),
@@ -1316,6 +551,7 @@ impl Render for DiffViewer {
                                 div()
                                     .flex_1()
                                     .bg(cx.theme().colors().surface_background)
+                                    .relative()
                                     .child(self.render_connectors(cx)),
                             ),
                     )
